pyproject.toml
[project]
name = "data-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "pooltool-billiards",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
pooltool-billiards = { path = "pooltool", editable = true }
